Refactor: [utils] 将 session_manager.py 移入 /utils 目录
